Update magAverage.R
Minor changes to avoid using indices in any calculations, for example ma1 instead of meanFlow. Good catch on not re-calculating means, I just prefer to use the variables. During my re-write of this it annoyed me to no end when indices were used in calculations because I kept having to refer back to the documentation or earlier in the code to remember what it was.